Moved parameter structs from VehicleConnection to ParameterServer
